backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
